+++ src/main/java/com/mobble/mobbleserver/application/club/core/service/ClubQueryService.java
@@ -1,10 +1,14 @@
 package com.mobble.mobbleserver.application.club.core.service;
 
+import com.mobble.mobbleserver.application.club.ageGroup.port.required.AgeGroupReadPort;
 import com.mobble.mobbleserver.application.club.core.port.provided.ClubQueryPort;
 import com.mobble.mobbleserver.application.club.core.port.required.ClubReadPort;
 import com.mobble.mobbleserver.application.clubMember.port.required.ClubMemberReadPort;
 import com.mobble.mobbleserver.application.ground.required.GroundReadPort;
 import com.mobble.mobbleserver.application.member.port.required.MemberReadPort;
+import com.mobble.mobbleserver.domain.address.Address;
+import com.mobble.mobbleserver.domain.club.ageGroup.AgeGroup;
+import com.mobble.mobbleserver.domain.club.ageGroup.AgeGroupType;
 import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.domain.clubMember.ClubMember;
 import com.mobble.mobbleserver.domain.clubMember.ClubMemberRole;
@@ -17,11 +21,7 @@ import com.mobble.mobbleserver.infrastructure.persistence.club.core.projection.C
 import com.mobble.mobbleserver.infrastructure.web.club.core.dto.request.ClubSearchRequestDto;
 import com.mobble.mobbleserver.infrastructure.web.club.core.dto.response.ClubResponseDto;
 import com.mobble.mobbleserver.infrastructure.web.club.core.dto.response.ClubSummaryDto;
-import com.mobble.mobbleserver.domain.address.Address;
 import com.mobble.mobbleserver.infrastructure.web.ground.dto.response.GroundResponseDto;
-import com.mobble.mobbleserver.refactor.club.ageGroup.entity.AgeGroup;
-import com.mobble.mobbleserver.refactor.club.ageGroup.entity.AgeGroupType;
-import com.mobble.mobbleserver.refactor.club.ageGroup.repository.AgeGroupRepository;
 import com.mobble.mobbleserver.refactor.club.clubGround.repository.ClubGroundRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -39,8 +39,8 @@ public class ClubQueryService implements ClubQueryPort {
     private final MemberReadPort memberReadPort;
     private final ClubMemberReadPort clubMemberReadPort;
     private final GroundReadPort groundReadPort;
+    private final AgeGroupReadPort ageGroupReadPort;
 
-    private final AgeGroupRepository ageGroupRepository;
     private final ClubGroundRepository clubGroundRepository;
 
     @Override
@@ -79,7 +79,7 @@ public class ClubQueryService implements ClubQueryPort {
     }
 
     private ClubResponseDto buildClubResponse(Club club, Member member, String leaderName) {
-        List<AgeGroupType> ageGroupList = ageGroupRepository.findByClubId(club.getId()).stream()
+        List<AgeGroupType> ageGroupList = ageGroupReadPort.findByClubId(club.getId()).stream()
                 .map(AgeGroup::getAgeGroupType)
                 .toList();
